refactor: update deviceType filename
