+++ .github/whitelist-capital-letters.txt
@@ -40,7 +40,6 @@
 ./libs/common/src/misc/serviceUtils.spec.ts
 ./libs/common/src/enums/kdfType.ts
 ./libs/common/src/enums/fileUploadType.ts
-./libs/common/src/enums/clientType.ts
 ./libs/common/src/enums/encryptedExportType.ts
 ./libs/common/src/enums/linkedIdType.ts
 ./libs/common/src/enums/encryptionType.ts
@@ -51,7 +50,6 @@
 ./libs/common/src/enums/eventType.ts
 ./libs/common/src/enums/hashPurpose.ts
 ./libs/common/src/enums/uriMatchType.ts
-./libs/common/src/enums/deviceType.ts
 ./libs/common/src/enums/secureNoteType.ts
 ./libs/common/src/enums/storageLocation.ts
 ./libs/common/src/enums/themeType.ts
