+++ libs/common/src/enums/device-type.ts
